[TECG-297]-Last Activity Date on Card: Code Review.
