+++ src/main/java/com/ciandt/techgallery/persistence/model/Technology.java
@@ -1,13 +1,19 @@
 package com.ciandt.techgallery.persistence.model;
 
-import java.util.Date;
-
-import com.ciandt.techgallery.service.transformer.TechnologyTransformer;
 import com.google.api.server.spi.config.ApiTransformer;
+
 import com.googlecode.objectify.annotation.Entity;
 import com.googlecode.objectify.annotation.Id;
 import com.googlecode.objectify.annotation.Index;
 import com.googlecode.objectify.annotation.Unindex;
+
+import com.ciandt.techgallery.service.enums.TechnologyOrderOptionEnum;
+import com.ciandt.techgallery.service.transformer.TechnologyTransformer;
+
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.Date;
+import java.util.List;
 
 /**
  * Technology entity.
@@ -93,73 +99,6 @@ public class Technology extends BaseEntity<String> {
 
   @Index
   private Boolean active;
-
-  /*
-   * Methods --------------------------------------------
-   */
-  /**
-   * Add 1 to the positive recomndations counter.
-   */
-  public void addPositiveRecommendationsCounter() {
-    positiveRecommendationsCounter++;
-  }
-
-  /**
-   * Remove 1 to the positive recomndations counter.
-   */
-  public void removePositiveRecommendationsCounter() {
-    if (positiveRecommendationsCounter > 0) {
-      positiveRecommendationsCounter--;
-    } else {
-      positiveRecommendationsCounter = 0;
-    }
-  }
-
-  /**
-   * Add 1 to the negative recomndations counter.
-   */
-  public void addNegativeRecommendationsCounter() {
-    negativeRecommendationsCounter++;
-  }
-
-  /**
-   * Remove 1 to the negative recomndations counter.
-   */
-  public void removeNegativeRecommendationsCounter() {
-    if (negativeRecommendationsCounter > 0) {
-      negativeRecommendationsCounter--;
-    } else {
-      negativeRecommendationsCounter = 0;
-    }
-  }
-
-  /**
-   * Add 1 to the commentary counter.
-   */
-  public void addCommentariesCounter() {
-    commentariesCounter++;
-  }
-
-  /**
-   * Remove 1 to the commentary counter.
-   */
-  public void removeCommentariesCounter() {
-    if (commentariesCounter > 0) {
-      commentariesCounter--;
-    } else {
-      commentariesCounter = 0;
-    }
-  }
-
-  /**
-   * Initialize the technology counters.
-   */
-  public void initCounters() {
-    commentariesCounter = 0;
-    endorsersCounter = 0;
-    negativeRecommendationsCounter = 0;
-    positiveRecommendationsCounter = 0;
-  }
 
   /*
    * Getter's and Setter's --------------------------------------------
@@ -304,5 +243,133 @@ public class Technology extends BaseEntity<String> {
 
   public void setLastActivityUser(String lastActivityUser) {
     this.lastActivityUser = lastActivityUser;
+  }
+
+  /*
+   * Methods --------------------------------------------
+   */
+  /**
+   * Add 1 to the positive recomndations counter.
+   */
+  public void addPositiveRecommendationsCounter() {
+    positiveRecommendationsCounter++;
+  }
+
+  /**
+   * Remove 1 to the positive recomndations counter.
+   */
+  public void removePositiveRecommendationsCounter() {
+    if (positiveRecommendationsCounter > 0) {
+      positiveRecommendationsCounter--;
+    } else {
+      positiveRecommendationsCounter = 0;
+    }
+  }
+
+  /**
+   * Add 1 to the negative recomndations counter.
+   */
+  public void addNegativeRecommendationsCounter() {
+    negativeRecommendationsCounter++;
+  }
+
+  /**
+   * Remove 1 to the negative recomndations counter.
+   */
+  public void removeNegativeRecommendationsCounter() {
+    if (negativeRecommendationsCounter > 0) {
+      negativeRecommendationsCounter--;
+    } else {
+      negativeRecommendationsCounter = 0;
+    }
+  }
+
+  /**
+   * Add 1 to the commentary counter.
+   */
+  public void addCommentariesCounter() {
+    commentariesCounter++;
+  }
+
+  /**
+   * Remove 1 to the commentary counter.
+   */
+  public void removeCommentariesCounter() {
+    if (commentariesCounter > 0) {
+      commentariesCounter--;
+    } else {
+      commentariesCounter = 0;
+    }
+  }
+
+  /**
+   * Initialize the technology counters.
+   */
+  public void initCounters() {
+    commentariesCounter = 0;
+    endorsersCounter = 0;
+    negativeRecommendationsCounter = 0;
+    positiveRecommendationsCounter = 0;
+  }
+
+  public static void sortTechnologiesDefault(List<Technology> techEntities) {
+    Collections.sort(techEntities, new Comparator<Technology>() {
+      @Override
+      public int compare(Technology counter1, Technology counter2) {
+        return counter2.getLastActivity().compareTo(counter1.getLastActivity());
+      }
+    });
+  }
+
+  public static List<Technology> sortTechnologies(List<Technology> techList,
+      TechnologyOrderOptionEnum orderBy) {
+    switch (orderBy) {
+      case POSITIVE_RECOMMENDATION_AMOUNT:
+        Collections.sort(techList, new Comparator<Technology>() {
+          @Override
+          public int compare(Technology counter1, Technology counter2) {
+            return Integer.compare(counter2.getPositiveRecommendationsCounter(),
+                counter1.getPositiveRecommendationsCounter());
+          }
+        });
+        break;
+      case NEGATIVE_RECOMMENDATION_AMOUNT:
+        Collections.sort(techList, new Comparator<Technology>() {
+          @Override
+          public int compare(Technology counter1, Technology counter2) {
+            return Integer.compare(counter2.getNegativeRecommendationsCounter(),
+                counter1.getNegativeRecommendationsCounter());
+          }
+        });
+        break;
+      case COMMENT_AMOUNT:
+        Collections.sort(techList, new Comparator<Technology>() {
+          @Override
+          public int compare(Technology counter1, Technology counter2) {
+            return Integer.compare(counter2.getCommentariesCounter(),
+                counter1.getCommentariesCounter());
+          }
+        });
+        break;
+      case ENDORSEMENT_AMOUNT:
+        Collections.sort(techList, new Comparator<Technology>() {
+          @Override
+          public int compare(Technology counter1, Technology counter2) {
+            return Integer.compare(counter2.getEndorsersCounter(), counter1.getEndorsersCounter());
+          }
+        });
+        break;
+      case APHABETIC:
+        Collections.sort(techList, new Comparator<Technology>() {
+          @Override
+          public int compare(Technology counter1, Technology counter2) {
+            return counter1.getName().compareTo(counter2.getName());
+          }
+        });
+        break;
+      default:
+        break;
+    }
+    return techList;
   }
 }
