+++ src/main/java/com/ciandt/techgallery/service/impl/TechnologyServiceImpl.java
@@ -29,8 +29,6 @@ import java.security.GeneralSecurityException;
 import java.text.Normalizer;
 import java.util.ArrayList;
 import java.util.Calendar;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 
@@ -143,15 +141,14 @@ public class TechnologyServiceImpl implements TechnologyService {
    * @throws BadRequestException in case a request with problem were made.
    */
   private void validateInformations(Technology technology) throws BadRequestException {
-    if (technology.getId() == null || technology.getId().equals("")) {
+    if (StringUtils.isBlank(technology.getId())) {
       throw new BadRequestException(ValidationMessageEnums.TECHNOLOGY_ID_CANNOT_BLANK.message());
-    } else if (technology.getName() == null || technology.getName().equals("")) {
+    } else if (StringUtils.isBlank(technology.getName())) {
       throw new BadRequestException(ValidationMessageEnums.TECHNOLOGY_NAME_CANNOT_BLANK.message());
-    } else
-      if (technology.getShortDescription() == null || technology.getShortDescription().isEmpty()) {
+    } else if (StringUtils.isBlank(technology.getShortDescription())) {
       throw new BadRequestException(
           ValidationMessageEnums.TECHNOLOGY_SHORT_DESCRIPTION_BLANK.message());
-    } else if (technology.getDescription() == null || technology.getDescription().equals("")) {
+    } else if (StringUtils.isBlank(technology.getDescription())) {
       throw new BadRequestException(ValidationMessageEnums.TECHNOLOGY_DESCRIPTION_BLANK.message());
     }
 
@@ -174,71 +171,10 @@ public class TechnologyServiceImpl implements TechnologyService {
       throw new NotFoundException(ValidationMessageEnums.NO_TECHNOLOGY_WAS_FOUND.message());
     } else {
       TechnologiesResponse response = new TechnologiesResponse();
-      sortTechnologiesDefault(techEntities);
+      Technology.sortTechnologiesDefault(techEntities);
       response.setTechnologies(techEntities);
       return response;
     }
-  }
-
-  private void sortTechnologiesDefault(List<Technology> techEntities) {
-    Collections.sort(techEntities, new Comparator<Technology>() {
-      @Override
-      public int compare(Technology counter1, Technology counter2) {
-        return counter2.getLastActivity().compareTo(counter1.getLastActivity());
-      }
-    });
-  }
-
-  private List<Technology> sortTechnologies(List<Technology> techList,
-      TechnologyOrderOptionEnum orderBy) {
-    switch (orderBy) {
-      case POSITIVE_RECOMMENDATION_AMOUNT:
-        Collections.sort(techList, new Comparator<Technology>() {
-          @Override
-          public int compare(Technology counter1, Technology counter2) {
-            return Integer.compare(counter2.getPositiveRecommendationsCounter(),
-                counter1.getPositiveRecommendationsCounter());
-          }
-        });
-        break;
-      case NEGATIVE_RECOMMENDATION_AMOUNT:
-        Collections.sort(techList, new Comparator<Technology>() {
-          @Override
-          public int compare(Technology counter1, Technology counter2) {
-            return Integer.compare(counter2.getNegativeRecommendationsCounter(),
-                counter1.getNegativeRecommendationsCounter());
-          }
-        });
-        break;
-      case COMMENT_AMOUNT:
-        Collections.sort(techList, new Comparator<Technology>() {
-          @Override
-          public int compare(Technology counter1, Technology counter2) {
-            return Integer.compare(counter2.getCommentariesCounter(),
-                counter1.getCommentariesCounter());
-          }
-        });
-        break;
-      case ENDORSEMENT_AMOUNT:
-        Collections.sort(techList, new Comparator<Technology>() {
-          @Override
-          public int compare(Technology counter1, Technology counter2) {
-            return Integer.compare(counter2.getEndorsersCounter(), counter1.getEndorsersCounter());
-          }
-        });
-        break;
-      case APHABETIC:
-        Collections.sort(techList, new Comparator<Technology>() {
-          @Override
-          public int compare(Technology counter1, Technology counter2) {
-            return counter1.getName().compareTo(counter2.getName());
-          }
-        });
-        break;
-      default:
-        break;
-    }
-    return techList;
   }
 
   /**
@@ -297,10 +233,10 @@ public class TechnologyServiceImpl implements TechnologyService {
       return new TechnologiesResponse();
     } else {
       if (techFilter.getOrderOptionIs() != null && !techFilter.getOrderOptionIs().isEmpty()) {
-        filteredList = sortTechnologies(filteredList,
+        filteredList = Technology.sortTechnologies(filteredList,
             TechnologyOrderOptionEnum.fromString(techFilter.getOrderOptionIs()));
       } else {
-        sortTechnologiesDefault(filteredList);
+        Technology.sortTechnologiesDefault(filteredList);
       }
       TechnologiesResponse response = new TechnologiesResponse();
       response.setTechnologies(filteredList);
